Refactor files and improve PropertyChanged event in ConfigBase
This significant reorganization involves the relocation of various Command and WiThrottle files to more relevant directories. Also added a check for changes in properties before triggering PropertyChanged or PropertyChanging events in ConfigBase, which efficiently optimizes event triggers. Finally, the type of "Speed" property in ICmdLocoSetSpeed interface was updated from byte to DCCSpeed for improved readability.