Added support for Supported and UnSupported Adapters and version for commands. Also automatically load and register commands at runtime.
This allows us to load the commands supported by a controller and once we create or assign an adapter to the controller, it will filter the commands that are valid. This is to support things like on the NCE systems where commands depend on the adapter in use and the version number.