Refactor codebase to use GUIDs instead of multiple types for IDs
Much of the codebase has been refactored to use GUIDs as IDs for entities instead of using different types depending on the entity. This change was made across multiple classes, causing deletion, addition, and modification of several code files. Additionally, some minor corrections to spelling/errors and changes to naming or structure of classes were made for improved clarity and organization.