Implement server message feature for WiThrottle
This commit primarily adds a new feature for WiThrottle server that includes a message system to communicate with clients. It also refactors related classes like WiThrottleCmdFactory and WiThrottleConnection. It introduces changes to the Execute method in Throttle command classes that now returns an instance of IServerMsg instead of a string. As a result, it enhances the tracing capabilities of the system.