Remove WiThrottle commands and messages
This commit removes all the WiThrottle command and message classes and adds new message classes to process incoming commands. This change was made to streamline the process for interpreting and executing these commands. The new system employs a system of command Objets which process their own commands and remove the need for multiple command classes.