Update ToString methods and improve hardware ID handling
The ToString methods in various classes have been updated to include the connection ID for easier debugging. A handling for duplicate hardware IDs was implemented in the CmdHardware class to re-use previous connections when there's a temporary disconnect and reconnect. Several other improvements and error handling cases have been included as well.