Implemented Layout State Management features and tests
Added new functionality for layout state management in the application, which provides a way to track and retrieve the state of given objects. This includes creation of new LayoutStateManager class and its associated StateAPI for RESTful interactions. State tests have also been added to verify the new functionality. Removed file "LayoutStateService.cs" as it was no longer being used.