Refactor loading mechanism for repositories
This commit updates the loading mechanism for the repositories in the RailwayManager. Previously, all the repositories were loaded on instantiation. This has now been refactored to accept a boolean parameter, which determines whether to load the repositories upon creation or not. The changes also include removal of redundant 'Save' method in 'IRailwayManager'.