Updated railway route configurations and improved route deactivation logic
The commit includes updates to the railway route configurations in the settings and layout files. Route IDs and keys have been swapped. A new property, DeActivateOnOverlap, has been added to the route entity. The route deactivation logic in CmdPanel.cs is updated to be more dynamic and only deactivate routes where there is a clash. This commit also adds three more turnouts to the turnout settings, improves the IP duplication detection logic in WiThrottleConnections.cs and adds initialization to some properties in WiThrottleConnection.cs.