Introduced WebApi integration and modified application start logic
Various changes were made to the application's start-up logic to include a WebApi integration. A new WebApi project was added along with associated configuration files and the necessary project references were added to the main application. Additionally, the WiThrottle logic was moved to a separate class from the main program.